requirements.txt
asgiref==3.9.1
Django==4.2.9
django-cors-headers==4.0.0
djangorestframework==3.16.1
package_name==0.1
pillow==11.3.0
python-dotenv==1.0.0
req==1.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
